+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Basic plotting function for sample results
+- Test for the plotting function (based on Covid model)
 
 ### Changed