+++ CHANGELOG.md
@@ -23,10 +23,12 @@ All notable changes to this project will be documented in this file.
 
 - Basic plotting function for sample results
 - Test for the plotting function (based on Covid model)
-- Test for general models that runs a fast sampling-based inference and visualizes the output
-- Test for the Test of general models
+- Users can now check the models they want to use with a basic functionality check and a quick inference check on artificial data. The inference check also calls the basic functionality check.
+- Test for the model checks.
 
 ### Changed
+
+- By default, each model is checked for basic functionality before attempting an inference run. However, this can be disabled using the argument check_model.
 
 ### Fixed