+++ CHANGELOG.md
@@ -23,7 +23,7 @@ All notable changes to this project will be documented in this file.
 
 - Basic plotting function for sample results
 - Users can now check the models they want to use with a basic functionality check and a quick inference check on artificial data.
-- The Docs now include code usage examples for many functions interesting to the user.
+- Added code usage examples to documentation.
 
 ### Changed